Terminado metodo que obtinee los movimientos del peon, y verfica si es posible el movimiento - con el codigo nuevo reestrcuturado

+++ Ajedrez/src/vector2d.h
@@ -1,9 +1,9 @@
 #pragma once
 class Vector2D {
 public:
-	double x, y;
+	int x, y;
 
 public:
-	Vector2D(double x, double y) : x(x), y(y) {};
+	Vector2D(int x, int y) : x(x), y(y) {};
 	Vector2D() = default;
 };
